requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
groq
pypdf
python-dotenv
langchain-groq
langchain-huggingface
torch
torchvision